+++ CHANGELOG.md
@@ -1,4 +1,27 @@
 # Changelog: [mojaloop/thirdparty-scheme-adapter](https://github.com/mojaloop/thirdparty-scheme-adapter)
+### [11.25.1](https://github.com/mojaloop/thirdparty-scheme-adapter/compare/v11.25.0...v11.25.1) (2021-03-12)
+
+
+### Maintenance
+
+* add request headers to accounts endpoint ([#60](https://github.com/mojaloop/thirdparty-scheme-adapter/issues/60)) ([b8b35b6](https://github.com/mojaloop/thirdparty-scheme-adapter/commit/b8b35b693bd2b13ef0a45e884cdc1a4eaf673c6d))
+
+## [11.25.0](https://github.com/mojaloop/thirdparty-scheme-adapter/compare/v11.24.3...v11.25.0) (2021-03-11)
+
+
+### Features
+
+* Add Inbound and Outbound accounts endpoints ([#56](https://github.com/mojaloop/thirdparty-scheme-adapter/issues/56)) ([2aa08f7](https://github.com/mojaloop/thirdparty-scheme-adapter/commit/2aa08f7891dadc1081152a26dca8c915558369e9))
+
+### [11.24.3](https://github.com/mojaloop/thirdparty-scheme-adapter/compare/v11.24.2...v11.24.3) (2021-03-11)
+
+### [11.24.2](https://github.com/mojaloop/thirdparty-scheme-adapter/compare/v11.24.1...v11.24.2) (2021-03-10)
+
+
+### CI/CD
+
+* add build-local dep to publish step ([#57](https://github.com/mojaloop/thirdparty-scheme-adapter/issues/57)) ([a1e3aee](https://github.com/mojaloop/thirdparty-scheme-adapter/commit/a1e3aee880f9e227a1f436db101fc883e9daeafa))
+
 ### [11.24.1](https://github.com/mojaloop/thirdparty-scheme-adapter/compare/v11.24.0...v11.24.1) (2021-03-09)